Add silx.gui.utils.async and move submitToQtMainThread there
