+++ silx/gui/utils/test/__init__.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 # /*##########################################################################
 #
-# Copyright (c) 2016-2018 European Synchrotron Radiation Facility
+# Copyright (c) 2018 European Synchrotron Radiation Facility
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -22,33 +22,24 @@
 # THE SOFTWARE.
 #
 # ###########################################################################*/
+"""silx.gui.utils tests"""
+
+
 __authors__ = ["T. Vincent"]
 __license__ = "MIT"
-__date__ = "28/11/2017"
+__date__ = "09/03/2018"
 
 
-from numpy.distutils.misc_util import Configuration
+import unittest
+
+from . import test_async
 
 
-def configuration(parent_package='', top_path=None):
-    config = Configuration('gui', parent_package, top_path)
-    config.add_subpackage('_glutils')
-    config.add_subpackage('qt')
-    config.add_subpackage('plot')
-    config.add_subpackage('fit')
-    config.add_subpackage('hdf5')
-    config.add_subpackage('widgets')
-    config.add_subpackage('test')
-    config.add_subpackage('plot3d')
-    config.add_subpackage('data')
-    config.add_subpackage('dialog')
-    config.add_subpackage('utils')
-    config.add_subpackage('utils.test')
+def suite():
+    """Test suite for module silx.image.test"""
+    test_suite = unittest.TestSuite()
+    test_suite.addTest(test_async.suite())
+    return test_suite
 
-    return config
-
-
-if __name__ == "__main__":
-    from numpy.distutils.core import setup
-
-    setup(configuration=configuration)
+if __name__ == '__main__':
+    unittest.main(defaultTest='suite')
